rename HTTPResponse.h
